+++ core/remote/src/main/kotlin/com/teammanduk/adego/core/remote/model/TmapReverseGeocodingResponse.kt
